Desktop/hybrid/_archive/old_dirs/v2_old/requirements.txt
# Hybrid v2 Bağımlılıkları
pygame-ce>=2.5.0
pygame_gui>=0.6.9
